+++ snout-boops/CHANGELOG.md
@@ -3,8 +3,19 @@
 ## Table of Contents <!-- omit in toc -->
 
 - [Unreleased](#unreleased)
+- [v0.1.0 - 2021-02-21](#v010---2021-02-21)
 
 ## Unreleased
+
+* name project
+* fix performance problems by using canvas rather than DOM elements
+* implement looping spike plot
+* implement looping spike graph
+* support resizing
+* break components and models into separate files
+* add monitoring and performance info
+
+## v0.1.0 - 2021-02-21
 
 * Removed some stock Create React App components and styles
 * Interface skeleton with a section for the spike map and the spike graph